feat(ServerConfig): 添加 ServerConfig 类型定义，优化配置管理 refactor(CopilotPreview, MainSection): 优化预览组件和主区域按钮样式，添加文件列表切换功能 refactor(MCPConfigContext): 移除冗余的 ServerConfig 定义，调整导入路径
