+++ package.json
@@ -23,6 +23,7 @@
     "clsx": "^2.1.1",
     "echarts-for-react": "^3.0.3",
     "eventsource": "^3.0.6",
+    "framer-motion": "^12.6.3",
     "lucide-react": "^0.475.0",
     "mermaid": "^11.6.0",
     "next": "15.2.0-canary.73",
